allow to customize building the squashfs image
Allow to configure a prefix for the cvmfs copy in the squashfs image (to emulate a default mount point) and also to merge one or more additional directories into the image. The latter is useful to create a OS base image plus cvmfs repository. Adjust the logging options to have more useful output with the single -v option. Change some of the print statements to more approprate logging statements.